Update detail handler
The detail handler now extends a Map
It also now includes all of the hardcoded values
