+++ api/src/api/context.ts
@@ -1,6 +1,5 @@
 import { DetailHandler } from "@/classes";
 import { BakalariParser, SSSVTParser } from "@/parser";
-import { classes, rooms, teachers } from "@/static";
 
 const details = new DetailHandler();
 
@@ -9,9 +8,6 @@ const context = {
     bakalariParser: new BakalariParser(details),
     sssvtParser: new SSSVTParser(details)
 };
-
-// Load all static details
-details.add(...classes, ...teachers, ...rooms);
 
 export default context;
 export type ApiContext = typeof context;
