+++ api/src/classes/details/index.ts
@@ -1,0 +1,2 @@
+export { Detail, DetailType, TeacherDetail, type AnyDetail } from "@/classes/details/details";
+export { DetailHandler } from "@/classes/details/handler";
